Set xml field names to match the swagger model
This change makes the XML annotations for property names reference the
swagger definition names rather than the generated class names.

If class names are used in the Java template then the XML produced by
other implementations may not be compatible. i.e. if your Swagger
model has an entity My-Model-Entity this will currently be serialized
as MyModelEntity (the class name) in the Java generated XML but other
languages which have different naming conventions would serialize in
different ways.

To make the XML language independent it should reflect the Swagger
model naming. That way if you want to use kebab case *my-model-entity*,
or snake case *my_model_entity* or any other naming convention you can
do this in your swagger model and know the XML is langage independent.

If you really want your model naming convention to be different to
what is in your XML (I don't know why you would) you can override this
with the xml.name tags.

Using code generated with this fix will be a breaking change for
clients if...
* They are using XML
AND
* The swagger property names are not Java style myObject